requirements.txt
pandas==2.0.3
requests==2.31.0
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
numpy==1.24.3
scipy==1.11.1
jupyter==1.0.0
python-dotenv==1.0.0
loguru==0.7.0